Use memory block to store data instead of node
Previously, we use node to store data for containers, which is
convenient, but with low efficiency, now we're using sequential memory
blocks to store data.  When space is not enough, containers automatically
expand it's data buff by 2 times. For JQueue, we use circular queue
technique to make use of memory.